v1.1.1 - Premium edition is up for grabs now! - Update AGP version - Update dependencies' version - Add screenshots & update README.md - Bug fixed

+++ app/src/main/java/top/xjunz/tasker/premium/PremiumMixin.kt
@@ -6,6 +6,7 @@ package top.xjunz.tasker.premium
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import top.xjunz.tasker.upForGrabs
 import x.f
 import java.io.ByteArrayInputStream
 import java.io.File
@@ -80,12 +81,16 @@ object PremiumMixin {
     }
 
     fun loadPremiumFromFileSafely() {
-        val file = File(premiumContextStoragePath)
-        if (file.exists()) {
-            runCatching {
-                deserializeFromString(file.readText())
-            }.onFailure {
-                file.delete()
+        if (upForGrabs) {
+            PREMIUM_CONTEXT = PremiumContext()
+        } else {
+            val file = File(premiumContextStoragePath)
+            if (file.exists()) {
+                runCatching {
+                    deserializeFromString(file.readText())
+                }.onFailure {
+                    file.delete()
+                }
             }
         }
     }
